Keep simulation CLI cleanup below warning thresholds
The simulation runner now stays as a small dispatcher while preset execution, swarm execution, and CLI/config parsing live in focused modules. The split preserves preset and swarm behavior, keeps invariant and no-worker-thread guards pointed at the real call sites, and lowers the maintenance ledger by one file-bloat warning.

Constraint: The public command remains `npx tsx scripts/run-simulation.ts` for both preset and swarm modes.
Rejected: Leaving the extracted modules untracked because scripts are ignored by default | CI would only receive the dispatcher and fail imports.
Confidence: high
Scope-risk: narrow
Directive: Keep future run-simulation changes in the focused modules and update the source-scan guards when call sites move.
Tested: npm.cmd test -- --watchAll=false --runTestsByPath src/simulation/__tests__/cli-invariant-wiring.test.ts src/simulation/__tests__/no-worker-threads.test.ts --runInBand; npx.cmd tsx scripts/run-simulation.ts --help; npx.cmd tsx scripts/run-simulation.ts --count=1 --seed=12345 --preset=light --output=%TEMP%\mekstation-run-simulation-smoke-2; npx.cmd tsx scripts/run-simulation.ts --config=scripts/swarm-configs/duel-3kbv-temperate.json --runs=1 --seed=42 --output=%TEMP%\mekstation-swarm-smoke-output-2.json; npm.cmd run maintain:scan -- --category=file-bloat --scope=scripts/run-simulation.ts,scripts/run-simulation-cli.ts,scripts/run-simulation-preset.ts,scripts/run-simulation-swarm.ts --json; npm.cmd run verify:qc:maintenance; npx.cmd oxfmt --check docs\qc\maintenance-warning-ledger.json scripts\run-simulation.ts scripts\run-simulation-cli.ts scripts\run-simulation-preset.ts scripts\run-simulation-swarm.ts src\simulation\__tests__\cli-invariant-wiring.test.ts src\simulation\__tests__\no-worker-threads.test.ts; git diff --cached --check
Not-tested: Full verify:qc was not rerun for this narrow maintenance slice; PR CI should run the full matrix.
